:sparkles: me error report and add subscription status to user out
