+++ src/app/routes/account.py
@@ -1,12 +1,16 @@
+from typing import Optional, Tuple
+
 from fastapi import APIRouter, Depends, HTTPException, Query, status
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy.orm import Session
- 
-from app.routes.dependencies import get_current_active_user
+
+from app.routes.dependencies import get_current_active_user, get_pagination_params, get_sort_by_params
 from app.crud import user_crud
+from app.crud.error_reports import error_report_crud
 from app.database.db import get_db
 from app.log import get_logger
 from app.models import User
+from app.schemas.error_report import ErrorReportOutPaginated
 from app.schemas.user import UserOut, UserUpdateOwn
 from app.schemas.auth import EmailChangeRequest
 from app.security import get_password_hash, verify_password
@@ -98,6 +102,49 @@ def update_current_active_user(
             detail=f"Couldn't update current active user. Error: {str(e)}",
         ) from e
     return user
+
+@router.get(
+    "/error-reports",
+    response_model=Optional[ErrorReportOutPaginated],
+    status_code=status.HTTP_200_OK,
+)
+def fetch_my_error_reports(
+    db: Session = Depends(get_db),
+    pagination_params: Tuple[int, int] = Depends(get_pagination_params),
+    orderby_params: Tuple[str, bool] = Depends(get_sort_by_params),
+    active_user: User = Depends(get_current_active_user),
+) -> Optional[ErrorReportOutPaginated]:
+    """
+    Fetch the error reports created by the current user.
+
+    Parameters:
+        db (Session): The database session.
+        pagination_params (Tuple[int, int]): The pagination parameters (skip, limit).
+        orderby_params (Tuple[str, bool]): The order by parameters (sortby, descending).
+        active_user (User): The current active user.
+
+    Returns:
+        Optional[ErrorReportOutPaginated]: The current user's error reports with pagination datas.
+    """
+    page, size = pagination_params
+    sortby, descending = orderby_params
+    error_reports, total = error_report_crud.get_many(
+        db,
+        skip=page,
+        limit=size,
+        order_by=sortby,
+        descending=descending,
+        created_by=active_user.id,
+    )
+    pages = (total + size - 1) // size
+    return {
+        "items": error_reports,
+        "total": total,
+        "page": page,
+        "size": size,
+        "pages": pages
+    }
+
 
 @router.patch("/email", status_code=status.HTTP_200_OK)
 def request_email_change(
